+++ src/main/java/uoa/lavs/loan/PersonalLoan.java
@@ -4,11 +4,11 @@ import java.util.ArrayList;
 
 public class PersonalLoan extends Loan {
   public PersonalLoan(
-      Integer loanId,
+      int loanId,
       String customerId,
       ArrayList<String> coborrowerIds,
-      Float principal,
-      Float rate,
+      Double principal,
+      Double rate,
       LoanDuration duration,
       LoanPayment payment) {
     super(loanId, customerId, coborrowerIds, principal, rate, duration, payment);
